Feat: style

+++ database/migrations/2024_11_14_094755_create_bookings_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->date('order_date');
             $table->date('check_in');
             $table->date('check_out');
-            $table->decimal('discount', 5, 2); 
+            $table->integer('discount'); 
             $table->text('notes')->nullable(); 
             $table->foreignId('room_id')->constrained('rooms')->onDelete('cascade');     
             $table->enum('status',['Booked','Pending','Refund','Cancelled']);
